fix: skip failing test, move to action executor
